Use a faster upload strategy
copying the source from local to a new location on the remote, is a safe
stragegy. But, it's slow if you are uploading anything of significant
size.

Rather than always upload the whole content:

* Copy the _existing_ version as a start point for the new deployment
* syncronize local with the new version
* Then (as before) link to the new version

This means that rsync only needs to send diffs and new files, not the
whole site content - The end result is the same, but transfer times are
less.